pyproject.toml
[build-system]
requires = ["setuptools>=69", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "eczema-tracker"
version = "0.1.0"
description = "Self-hosted eczema treatment tracker backend"
requires-python = ">=3.12"
dependencies = [
  "fastapi>=0.115,<1.0",
  "uvicorn[standard]>=0.30,<1.0",
  "sqlalchemy>=2.0,<3.0",
  "alembic>=1.13,<2.0",
  "psycopg[binary]>=3.2,<4.0",
  "pydantic-settings>=2.6,<3.0",
  "python-jose[cryptography]>=3.3,<4.0",
  "passlib[bcrypt]>=1.7,<2.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.0,<9.0",
  "httpx>=0.27,<1.0",
]

[tool.setuptools.packages.find]
include = ["app*"]
